split up utils into seperate files and move utils functionality from lib folder
