+++ src/utils/calculateAverage.ts
@@ -1,0 +1,4 @@
+export const calculateAverage = (arr: number[]): number => {
+  const average = arr.reduce((a, b) => a + b, 0) / arr.length;
+  return parseFloat(average.toFixed(1));
+};
